refactor: useHistory hook
