+++ src/hooks/persisted/index.ts
@@ -1,3 +1,4 @@
 export { useTheme } from './useTheme';
 export { useUpdates, useLastUpdate } from './useUpdates';
 export { default as useCategories } from './useCategories';
+export { default as useHistory } from './useHistory';
