+++ src/hooks/persisted/useHistory.ts
@@ -1,31 +1,24 @@
 import { useCallback, useState } from 'react';
 import { useFocusEffect } from '@react-navigation/native';
 
-import { History } from '../database/types';
+import { History } from '@database/types';
 
 import {
   deleteAllHistory,
   deleteChapterHistory,
   getHistoryFromDb,
-} from '../database/queries/HistoryQueries';
+} from '@database/queries/HistoryQueries';
 
 const useHistory = () => {
   const [isLoading, setIsLoading] = useState(true);
   const [history, setHistory] = useState<History[]>([]);
   const [error, setError] = useState<string>();
 
-  const getHistory = async () => {
-    try {
-      const res = await getHistoryFromDb();
-      setHistory(res);
-    } catch (err) {
-      if (err instanceof Error) {
-        setError(err.message);
-      }
-    } finally {
-      setIsLoading(false);
-    }
-  };
+  const getHistory = () =>
+    getHistoryFromDb()
+      .then(res => setHistory(res))
+      .catch((err: Error) => setError(err.message))
+      .finally(() => setIsLoading(false));
 
   const clearAllHistory = () => {
     deleteAllHistory();
